requirements.txt
flask
flask-cors
pillow
pyyaml
requests
pygame
